refactor: Improve formatting and structure of migration for team_id and created_by columns

+++ apps/api/database/migrations/16_add_team_id_to_opportunities.ts
@@ -1,4 +1,3 @@
-
 import BaseSchema from '@ioc:Adonis/Lucid/Schema'
 
 export default class extends BaseSchema {
@@ -6,17 +5,33 @@ export default class extends BaseSchema {
 
   public async up() {
     this.schema.alterTable(this.tableName, (table) => {
-      table.integer('team_id').unsigned().nullable().references('id').inTable('teams').onDelete('SET NULL')
-      table.integer('created_by').unsigned().nullable().references('id').inTable('users').onDelete('SET NULL')
+      table
+        .integer('team_id')
+        .unsigned()
+        .nullable()
+        .references('id')
+        .inTable('teams')
+        .onDelete('SET NULL')
+      table
+        .integer('created_by')
+        .unsigned()
+        .nullable()
+        .references('id')
+        .inTable('users')
+        .onDelete('SET NULL')
       table.string('checkin_code').nullable()
     })
   }
 
   public async down() {
-    this.schema.alterTable(this.tableName, (table) => {
-      table.dropColumn('team_id')
-      table.dropColumn('created_by')
-      table.dropColumn('checkin_code')
-    })
+    if (await this.schema.hasTable(this.tableName)) {
+      this.schema.alterTable(this.tableName, (table) => {
+        table.dropForeign(['team_id'])
+        table.dropForeign(['created_by'])
+        table.dropColumn('team_id')
+        table.dropColumn('created_by')
+        table.dropColumn('checkin_code')
+      })
+    }
   }
 }
